calculator

+++ calculator.py
@@ -2,8 +2,8 @@ num1 = input("Enter a Number: ")
 num2 = input("Enter a Number: ")
 
 # Convert input values to numbers
-num1 = int(num1)
-num2 = int(num2)
+num1 = float(num1)
+num2 = float(num2)
 
 result = num1 + num2
 print(result)
